fix: correct OAEPparams namespace and derive MGF map single-source
- OAEPparams now emits as e:OAEPparams (xenc namespace), not bare
  OAEPparams (inherited xmldsig namespace). Schema-validating peers
  require it in xenc per EncryptionMethodType.
- Test updated to assert namespaceURI and element ordering via DOM.
- MGF_ALGORITHMS now defined once in lib/mgf-algorithms.js.
  Template derives short→URI map, ensuring sha1 emits xmlenc11#mgf1sha1
  (not the legacy alias). Drift now caught: unmapped short names throw
  instead of emitting bare Algorithm="name".
- Removed || keyEncryptionMgf fallback that silently serialized
  non-URIs. Unmapped values now fail loudly.
