+++ lib/xmlenc.js
@@ -3,6 +3,7 @@ var xmldom  = require('@xmldom/xmldom');
 var xpath   = require('xpath');
 var utils   = require('./utils');
 var oaep    = require('./oaep');
+var { MGF_ALGORITHMS, MGF_SHORT_NAMES } = require('./mgf-algorithms');
 
 const insecureAlgorithms = [
   //https://www.w3.org/TR/xmlenc-core1/#rsav15note
@@ -13,19 +14,6 @@ const insecureAlgorithms = [
   'http://www.w3.org/2001/04/xmlenc#aes256-cbc',
   'http://www.w3.org/2001/04/xmlenc#aes128-cbc',
 ];
-
-// XML-Enc 1.1 5.5.2. The normative list uses the xmlenc11#mgf1* URIs; the
-// xmlenc#MGF1withSHA1 spelling is accepted on decrypt because Example 33 in
-// that same section uses it and implementations copied it.
-const MGF_ALGORITHMS = Object.assign(Object.create(null), {
-  'http://www.w3.org/2009/xmlenc11#mgf1sha1': 'sha1',
-  'http://www.w3.org/2009/xmlenc11#mgf1sha224': 'sha224',
-  'http://www.w3.org/2009/xmlenc11#mgf1sha256': 'sha256',
-  'http://www.w3.org/2009/xmlenc11#mgf1sha384': 'sha384',
-  'http://www.w3.org/2009/xmlenc11#mgf1sha512': 'sha512',
-  'http://www.w3.org/2001/04/xmlenc#MGF1withSHA1': 'sha1'
-});
-const MGF_SHORT_NAMES = Object.values(MGF_ALGORITHMS);
 
 function encryptKeyInfoWithScheme(symmetricKey, options, padding, mgf1Hash, callback) {
   const symmetricKeyBuffer = Buffer.isBuffer(symmetricKey) ? symmetricKey : Buffer.from(symmetricKey, 'utf-8');
